+++ example/lib/main.dart
@@ -6,7 +6,7 @@ import 'package:flutter_modular/flutter_modular.dart';
 void main() async {
   try {
     WidgetsFlutterBinding.ensureInitialized();
-    await Meteor.connect('ws://192.168.27.113:3000/websocket');
+    await Meteor.connect('ws://10.0.2.2:3000/websocket');
     runApp(ModularApp(module: AppModule()));
   } on MeteorError catch (error) {
     print(error);